requirements.txt
fastapi[standard]
ultralytics
Pillow